# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ZIAY API
  version: 1.0.0
  description: |
    Comercio Conversacional + Atribución Inteligente.

    Multi-tenant conversational-commerce platform with agent-driven checkout
    (AP2 / UCP / ACP / MCP protocol trinity), ad attribution, compliance
    (Ley 1581 / Ley 2573 / Ley 1480 / DIAN), and a wallet/commission layer
    for traffickers + marketplaces.

    All tenant-scoped endpoints require a session cookie (`sessionAuth`)
    unless marked otherwise. UCP / ACP / webhook endpoints use `bearerAuth`
    or are public.

  contact:
    email: api@ziay.co
  license:
    name: Proprietary
servers:
  - url: https://ziay.co
    description: Production
  - url: http://localhost:3000
    description: Development
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth v4 session cookie (set by /api/auth/[...nextauth]/callback/credentials).
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed W3C Verifiable Credential or bearer token (UCP/AP2/ACP).
    cronAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Bearer $CRON_SECRET — shared secret for /api/compliance/retention/cron.'
    webhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: HMAC-SHA256 signature from the gateway (Meta/Stripe/Wompi/etc.).
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
...
  parameters:
    TenantId:
      name: tenantId
      in: query
      required: true
      schema:
        type: string
      description: Tenant slug (e.g. "saramantha").
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor for cursor-based pagination (id of the last item on the previous page).
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
      description: Page size (max 100).
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Tenant access denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - sessionAuth: []
paths:
...
  /api/orders:
    get:
      summary: List orders (cursor-paginated)
      description: Filter by status / mode / search query. Cursor-based pagination (max 100).
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: status
          in: query
          schema:
            type: string
        - name: mode
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated orders + nextCursor + hasMore.
  /api/orders/{id}:
    patch:
      summary: Update order
      description: Update status / payment status. Optional `event` is recorded atomically in a transaction.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                paymentStatus:
                  type: string
                paidAt:
                  type: string
                  format: date-time
                paymentGateway:
                  type: string
                paymentRef:
                  type: string
                event:
                  type: string
                note:
                  type: string
      responses:
        '200':
          description: Updated order.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/channels:
    get:
      summary: List channels
      description: List channels for a tenant. Tokens are masked (hasToken flags only).
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Channel list with masked credentials.
    post:
      summary: Create channel
      description: Add a new channel (WhatsApp / Messenger / Instagram / Telegram). Validates per-type required fields.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tenantId:
                  type: string
                type:
                  type: string
                  enum: [whatsapp, messenger, instagram, telegram]
                name:
                  type: string
                displayName:
                  type: string
      responses:
        '201':
          description: Created channel.
        '400':
          $ref: '#/components/responses/BadRequest'
    patch:
      summary: Update channel
      description: Update a channel (e.g., rotate credentials). Tenant ownership verified.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                channelId:
                  type: string
      responses:
        '200':
          description: Updated channel.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Deactivate channel (soft delete)
      description: Soft-delete a channel — preserves conversation history.
      security:
        - sessionAuth: []
      parameters:
        - name: channelId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deactivated channel id.
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/integrations/credentials:
    get:
      summary: List integration credentials (masked)
      description: List credential Setting rows (`cred::*` keys) with masked values.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Masked credential list.
    post:
      summary: Create / update integration credential
      description: Upsert a `cred::*` Setting row.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Upserted credential.
    put:
      summary: Replace integration credential
      description: Replace a `cred::*` Setting row.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Replaced credential.
    delete:
      summary: Delete integration credential
      description: Delete a `cred::*` Setting row.
      security:
        - sessionAuth: []
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted credential.

  # ─────────────────────────────────────────────────────────────────────
  # 7. ADS, TRAFFICKER, ATTRIBUTION
  # ─────────────────────────────────────────────────────────────────────
...
  /api/payments/config:
    get:
      summary: List payment config
      description: Payment strategy per channel + non-credential global thresholds.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: '{ channels, global }'
    patch:
      summary: Update payment config
      description: Update a channel's payment strategy + whitelisted global Setting keys.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated channel + changed setting keys.
...
  /api/novedades/{id}:
    get:
      summary: Fetch novedad detail
      description: Single case with events + assignee + customer.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Case detail.
    patch:
      summary: Update novedad
      description: Update case fields (status, priority, assignment).
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated case.
...